selenium_tests/requirements.txt
selenium==4.18.1
pytest==8.0.2
webdriver-manager==4.0.1
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2
requests==2.31.0
python-dotenv==1.0.0 